e2e test for load neighbours
