--- workbase/test/e2e/DoubleClick.test.js
@@ -1,67 +1,0 @@
-const Application = require('spectron').Application;
-const assert = require('assert');
-const electronPath = require('electron'); // Require Electron from the binaries included in node_modules.
-const path = require('path');
-
-const sleep = time => new Promise(r => setTimeout(r, time));
-jest.setTimeout(15000);
-
-const app = new Application({
-  path: electronPath,
-  args: [path.join(__dirname, '../../dist/electron/main.js')],
-});
-
-beforeAll(async () => app.start());
-
-afterAll(async () => {
-  if (app && app.isRunning()) {
-    return app.stop();
-  }
-  return undefined;
-});
-
-describe('Favourite queries', () => {
-  test('initialize workbase', async () => {
-    const count = await app.client.getWindowCount();
-    assert.equal(count, 1);
-  });
-
-  test('select keyspace', async () => {
-    app.client.click('.keyspaces');
-    await app.client.waitUntilWindowLoaded();
-0
-    const keyspaceList = app.client.selectByAttribute('class', 'keyspaces-list');
-    assert.ok(keyspaceList);
-
-    assert.equal(await app.client.getText('.keyspaces'), 'keyspace');
-
-    app.client.click('#gene');
-
-    assert.equal(await app.client.getText('.keyspaces'), 'gene');
-  });
-
-  test('double click on entity', async () => {
-    app.client.click('.CodeMirror');
-
-    await sleep(1000);
-
-    app.client.keys('match $x isa person has attribute $y; $r($x, $z); $z isa entity; $r isa parentship; offset 0; limit 1; offset 0; get;');
-
-    await sleep(1000);
-
-    app.client.click('.run-btn');
-
-    await sleep(6000);
-
-    app.client.rightClick('#graph-div', 540, 470);
-    // app.client.leftClick('#graph-div', 540, 470);
-
-    await sleep(4000);
-  });
-
-  test('double click on attribute', async () => {
-  });
-
-  test('double click on relationship', async () => {
-  });
-});
